pyproject.toml
[project]
name = "bx-skills"
version = "0.1.0"
description = "TUI for installing AI coding assistant skills"
requires-python = ">=3.10"
dependencies = ["textual>=8.0.0", "rich-click>=1.8.0"]
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "bitranox", email = "bitranox@gmail.com" }]
keywords = ["cli", "skills", "tui", "textual", "ai-assistant"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "License :: OSI Approved :: MIT License",
  "Environment :: Console",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0",
  "pytest-asyncio>=0.25.0",
  "hypothesis>=6.151.9",
  "ruff>=0.15.1",
  "pyright[nodejs]>=1.1.408",
  "bandit>=1.9.3",
  "build>=1.4.0",
  "twine>=6.2.0",
  "codecov-cli>=11.2.6",
  "pip-audit>=2.10.0",
  "textual>=8.0.0",
  "import-linter>=2.10",
  "rtoml>=0.13.0",
  "httpx>=0.28.1",
  # Transitive CVE pins
  "jaraco-context>=6.1.0",       # CVE-2025-23949
  "python-multipart>=0.0.22",    # CVE-2025-24486
  "wheel>=0.46.3",               # CVE-2025-24049
  "pynacl>=1.6.2",               # CVE-2025-69277
  "urllib3>=2.6.3",              # CVE-2025-66471, CVE-2025-66418
  "virtualenv>=20.37.0",         # CVE-2025-22702
]

[project.scripts]
bx-skills = "bx_skills.cli:main"

[build-system]
requires = ["hatchling>=1.28.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bx_skills"]

[tool.hatch.build]
# Include all catalog data files (markdown, shell scripts, configs, images, etc.)
artifacts = [
  "src/bx_skills/catalog_skills/**/*",
  "src/bx_skills/catalog_commands/**/*",
]

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = [
  "src/bx_skills/catalog_skills",
]

[tool.ruff.lint]
select = [
  "E", "W",     # pycodestyle
  "F",           # pyflakes
  "I",           # isort
  "N",           # pep8-naming
  "UP",          # pyupgrade
  "B",           # bugbear
  "A",           # builtins shadowing
  "C4",          # comprehension simplification
  "SIM",         # simplification
  "TCH",         # type-checking imports
  "RUF",         # ruff-specific
  "G",           # logging format (catches f-string in logger)
  "T20",         # print statements
  "PTH",         # pathlib preference
  "PLC", "PLE",  # pylint conventions/errors
  "S",           # flake8-bandit security checks
  "PERF",        # performance anti-patterns
  "ERA",         # eradicate dead/commented-out code
  "FBT",         # boolean trap (boolean positional args)
  "PLR",         # pylint refactor
  "PLW",         # pylint warnings
]
ignore = [
  "RUF002",    # ambiguous unicode characters in docstrings (en-dash) - typography choice
  "RUF022",    # unsorted __all__ - kept in logical grouping order
  "PLC0415",   # deferred imports inside functions - circular import prevention
  "TC001",     # stdlib imports at runtime - needed, not just for type hints
  "TC002",     # third-party imports at runtime - needed, not just for type hints
  "TC003",     # stdlib imports - same as TC001
  "TC006",     # cast() type expressions - existing code uses unquoted
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# Notebooks: unused imports are often intentional for side effects
"notebooks/*.ipynb" = ["F401"]
# Tests: assert is standard, test fixtures have passwords, magic values are clearer inline, temp paths are fine
"tests/*.py" = ["T20", "S101", "S105", "S106", "S108", "FBT001", "FBT003", "PLR0913", "PLR2004"]
"src/bx_skills/cli.py" = ["FBT001", "FBT003", "PLR0913", "PLR0912", "PERF203"]

[tool.pyright]
typeCheckingMode = "standard"
extraPaths = [
  "src",
]
exclude = [
  "src/bx_skills/catalog_skills",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = ["--doctest-modules", "--ignore=src/bx_skills/catalog_skills"]
doctest_optionflags = ["ELLIPSIS", "NORMALIZE_WHITESPACE"]
pythonpath = ["src"]
markers = [
  "os_agnostic: test runs on every supported operating system",
  "os_windows: test exercises Windows-only behavior",
  "os_macos: test exercises macOS-only behavior",
  "os_posix: test exercises behavior requiring POSIX semantics",
  "os_linux: test exercises Linux-only behavior",
  "local_only: tests that require the local dev environment (external resources, script composition); skipped in CI",
  "integration: tests requiring external resources (SMTP, databases, APIs); run via 'bmk ti' or 'mktesti.sh'",
]

[tool.coverage.run]
branch = true
parallel = false
data_file = "/tmp/.coverage.bx_skills"
source = ["src/bx_skills"]
omit = [
  "src/bx_skills/catalog_skills/*",
]

[tool.coverage.report]
omit = [
  "tests/*",
  "src/bx_skills/catalog_skills/*",
]
fail_under = 80
show_missing = true

[tool.coverage.json]
output = "coverage.json"
pretty_print = true

[tool.bandit]
exclude_dirs = [
  "src/bx_skills/catalog_skills",
  "tests",
]

[tool.importlinter]
root_package = "bx_skills"
exclude_type_checking_imports = true

[tool.bashate]
max-line-length = 120
ignores = [
  "E003",
]

[tool.clean]
# Patterns to remove when running `make clean`
patterns = [
  "**/__pycache__",
  ".hypothesis",
  ".import_linter_cache",
  ".pytest_cache",
  ".ruff_cache",
  ".pyright",
  ".mypy_cache",
  ".tox",
  ".nox",
  ".eggs",
  "*.egg-info",
  "build",
  "dist",
  "htmlcov",
  ".coverage",
  "coverage.xml",
  "codecov.sh",
  ".cache",
  "result",
  ".venv",
]

[tool.git]
# Default git remote for push/release commands
default-remote = "origin"

[tool.scripts.test]
# Configuration for CI/CD test runner
pytest-verbosity = "-vv"
coverage-report-file = "coverage.xml"
src-path = "src"

[tool.pip-audit]
# Known vulnerabilities to ignore (environment-level packages, not project dependencies)
ignore-vulns = [
  "PYSEC-2022-42969",
  "PYSEC-2022-43012",  # setuptools 65.5.0 — env-only, not a project dependency
  "PYSEC-2025-49",     # setuptools 65.5.0 — env-only, not a project dependency
  "CVE-2024-6345",     # setuptools 65.5.0 — env-only, not a project dependency
  "CVE-2026-1703",
  "CVE-2026-25990",
  "CVE-2026-26007",
]
